refactor db getter logic
